applicationState_t totalFrameCount Uint64 to size_t

+++ main/applicationState_t/include/applicationState_t.h
@@ -6,5 +6,5 @@
 typedef struct {
     SDL_Window* window;
     SDL_Renderer* renderer;
-    Uint64 totalFrameCount;
+    size_t totalFrameCount;
 } applicationState_t;
